portfolio_django_2017/requirements.txt
# sudo pip3 install -r requirements.txt
Django==1.11
Pillow==4.2.1
requests==2.18.3
django-debug-toolbar==1.6
factory-boy==2.9.2
mock==2.0.0
